pyproject.toml
[project]
name = "local-rag-assistant"
version = "0.1.0"
description = "A local RAG assistant built with FastAPI, Ollama, LangChain, and Chroma."
readme = "doc/开发计划.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "chromadb>=0.5,<1.0",
    "fastapi>=0.115,<1.0",
    "httpx>=0.28,<1.0",
    "langchain>=0.3,<0.4",
    "langchain-chroma>=0.2,<0.3",
    "langchain-community>=0.3,<0.4",
    "langchain-ollama>=0.2,<0.3",
    "langchain-text-splitters>=0.3,<0.4",
    "pymupdf>=1.24,<2.0",
    "beautifulsoup4>=4.12,<5.0",
    "python-multipart>=0.0.12,<0.1",
    "pydantic-settings>=2.7,<3.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.14,<2.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.11,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
